refactor(main): extract lifespan and service registry into dedicated modules
Move startup/shutdown logic to core/lifespan.py and service initialization
to core/service_registry.py, reducing main.py from ~880 to ~50 lines.
